backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
requests
pydantic
speechrecognition
python-multipart
python-engineio
python-socketio
starlette
websockets 